Set current_worker_id using existing restore mechanism, remove call_wrapper
Summary:
This is just something I noticed while reading the code. This global state needs
to be set when the worker starts, but we already have a code path which does
this.

The call wrapper is executed for every job, but the worker ID is always the same
for a given worker, so we can just set it once when the worker starts.

With this change, the call wrapper is no longer used, so I delete it in this
diff.

Reviewed By: grievejia

Differential Revision: D56831116

fbshipit-source-id: aa9c57720811bae5eb9d94fb50e2e9bc4322b756