Remove unused handle abstraction
Summary:
OCaml represents a file descriptor differently in Windows. In Unix-like systems,
a file descriptor is just an int. In windows, it is a custom object wrapping a
win32 HANDLE value.

In order to pass a file descriptor to another process in Windows, we needed to
handle serialization / deserialization for the HANDLE value.

This code isn't used for Pyre because Pyre doesn't currently support Windows.
With OCaml 5 and domains, we will not have the need to pass file descriptors
across processes like this.

Reviewed By: grievejia

Differential Revision: D56831119

fbshipit-source-id: 0c89d8f855c3cb2a5a47a76c7a8a9aa9bad685b0