Fork workers instead of spawn, cleanup
Summary:
Back when we added Windows support to Flow, we changed Hack to spawn workers
instead of forking them. Compared to forking, spawning a process requires a
complicated process:

1. Register an entry point from the main process
2. Store the name of the entry point in the HH_SERVER_DAEMON env var
3. Serialize data that needs to be passed to the worker into a temp file and
store the path to that file in the HH_SERVER_DAEMON_PARAM env var
4. At startup, check to see if HH_SERVER_DAEMON is set, and if so deserialize
the temp file and jump to the entry function

Because we spawn a new process, we need to store any global state needed by the
worker, like logging state.

This diff reverts back to using fork. By using fork, avoid all of the steps
above. Any global state that is initialized before the fork will be initialized
in the forked process. We can skip over processes startup and jump directly to
the appropriate worker entry function instead.

As we migrate to OCaml 5, we will certainly not need this process spawning
infrastructure, so we can get rid of it now.

Reviewed By: grievejia

Differential Revision: D56831118

fbshipit-source-id: 4f22a590365ed3a7698cc8ee1f9ce16d1691ef7d