Remove unused socket channel mode for worker communication
Summary:
Pyre always uses a pair of pipes to communicate between the main process and
worker processes. There is no need to change this to sockets, so this diff gets
rid of the code to support socket communication.

As we move to OCaml 5 with domains for parallelism, we won't need any IPC for
workers, so it's fine to consolidate on pipes for now.

Reviewed By: grievejia

Differential Revision: D56831115

fbshipit-source-id: d5b17195c70a3ba448e0be2550237095d7427642