Don't let AttributeResolution produce bad locations in errors
Summary:
It turns out that the guts of attributeResolution were assuming that
arguments - which appear as types but have *optional* expressions
attached - would always have locations on the expressions. This was
not a type-safe assumption, because we would just fall back to
`Location.any`.

That was a huge mistake. It is *never* safe to allow `Location.any` to appear
somewhere where it can show up in a type error, and the "reason" data spit
out by attribute resolution code can propagate into a type error inside
of `typeCheck.ml`s `forward_call` function.

As a result we were throwing an error with location `-1:-1` on errors with
a mismatch on the second argument:
```
def f() -> None:
    d: dict[str, str] = {}
    d["k"], y = 5, 5
```
was producing type check results like this:
```
library.py:-1:-1 Incompatible parameter type [6]: In call `dict.__setitem__`, for 2nd positional argument, expected `str` but got `int`.
```

We want to stop using Location.any where it could leak into errors. The
safest change would be to stop using it everywhere.

Unfortunately doing that everywhere would be a time consuming change,
and we need a quick fix right now. The solution is to require a location
on the critical AttributeResolution calls. In places where we know that
the results won't ever be used for errors, we go ahead and pass Location.any
so that we don't have to do a big refactor of Pyre right now; inside of
the `forward_call` logic in `typeCheck` we *always* pass the location of
the underlying call so that if an argument doesn't have a location available
we will fall back to the call location in the error message.

**Why didn't our tests catch this?**

Our unit test coverage is fine, we actually *do* have code that produced
these junk locations. The problem is that our test *infra* just throws
away locations. If we used any style of test that allows us to see where
errors are, we would have caught this before it reached trunk.

NOTE: there are at least two good BE projects we could do here: one is to
try to reduce or eliminate the use of Location.any in as much production
code as we can (it's not that bad to allow it in unit tests, that can be
useful); the other is to change how our tests work so that it's obvious
where errors happen.

Reviewed By: grievejia

Differential Revision: D56863625

fbshipit-source-id: d013c15c9cdfbd780635a568955ff2b57d364aed