Document a regression in unawaitedAwaitableCheck
Summary:
I checked the CI results against trunk when using Pyre built against
the commit where I stop doing eager lowering of `__setitem__`, and realized
that we were throwing new unawaited awaitable false positives when people
assign awaitables into dictionaries.

I think the problem is that my logic in `forward_assign` is not correctly
reproducing the existing behavior for container assignments - my intuition
for the correct behavior wasn't great coming from type check, but what we
need to do is "transfer" the unawaited-ness "into" the container, which is
done by heavily special-cased logic of `forwar_call`.

To fix this I'm probably going to need a refactor that makes `forward_call`
operate at a higher level of abstraction than AST nodes, similar to the
refactor of `typeCheck.ml` that was needed to handle synthetic calls to
`__setitem__` there.

This test reproduces the problem without the parser change (which will allow
me to bugfix it and restack the parser change) by introducing a
multiple-target subscript assignment with awaitables in the values; my
previous cases all involved awaitables in the *left* hand side but not the
*right* hand side.

-------------

The next few diffs are going to be me wading through
`unawaitedAwaitableCheck` to build a good mental model of it, because
I don't think I can fix this issue without a pretty clear understanding of how
the reachability + ownership analysis actually works. As a result, I'll be doing
mostly inline-docs + refactoring commits for a bit before I get to fixing the bug
tracked by this test.

Reviewed By: yangdanny97

Differential Revision: D56938609

fbshipit-source-id: 0017bc40d6c42099cb1a0adaa170baf4744a4d17