Add a unit test for @dataclass_transform to demonstrate that "frozen" is not recognized as a keyword
Summary:
Context:
We want to support dataclass transforms as part of 3.12 support. Here, we focus on the dataclass_transform decorator.

Changes:
We add a unit test to show that the "frozen" keyword is not recognized.

              from typing import dataclass_transform, Any, TypeVar, Type
              from dataclasses import dataclass, field
              T = TypeVar("T")

              dataclass_transform()
              def custom_dataclass(cls: Type[T]) -> Type[T]:
                  return dataclass(cls, frozen=True)

              custom_dataclass
              class A:
                  x: int
              a = A(x=10)

The errors we get are:

`Invalid decoration [56]: Decorator `typing.dataclass_transform(...)` could not be called, because its type `unknown` is not callable.

Unexpected keyword [28]: Unexpected keyword argument `frozen` to call `dataclass``

NOTE:
The above code type checks on Pylance.

Reviewed By: stroxler

Differential Revision: D56739184

fbshipit-source-id: 3fd79030cc23c8e70590df27fac34a058efdf2c8