Differentiate between assignment statement with no RHS, and assignment statements with `...` as its RHS
Summary:
This diff makes it so that the parsers no longer automatically create `...` when an assignment has no RHS. This allows us to differentiate the two cases in various analyses (like uninitialized attrs as mentioned in the task).

I guess because assignments are so common, this ended up affecting a ton of other things.

eager_dog_tangle

As with before I tried to update the existing logic to treat `None` the same as `...`, creating the latter as necessary & marking with TODO when it seemed too complicated to refactor.

I'll handle those in followup diffs, if anyone has opinions on how the remaining cases should be handled, please LMK.

Reviewed By: stroxler

Differential Revision: D56715827

fbshipit-source-id: 46149bb80ede4a596a49d6d00db5c25dbff75d70