Use explicit `__getitem__` match in classHierarchyEnvironment
Summary:
I'm preparing to push for a new `Ast.Subscript` node, and to get rid of
`__getitem__` lowering from Pyre's parser(s).

To do this, it will be much much easier if as many uses of subscript as possible
are *explicitly* matching on `__getitem__` so that a search-and-replace
can find them instead of me having to rely on integration tests to find places
where we assumed all calls are `__getitem__` without putting an explicit marker
in the code.

I'm working on converting all the implicit (or otherwise
not-searchable-directly-by-callsite) uses of `__getitem__` with explicit
match statements that will be straightforward to just cut over. This diff
is fixing an implicit use in `classHierarchyEnvironment` where we were assuming
any calls inside of the parent class list of a class definition are in fact
`__getitem__` calls on generic parents. Let's just make this explicit.

Reviewed By: grievejia

Differential Revision: D57002336

fbshipit-source-id: da813a01fa004ffc41272c335a8b76d4ac00e226