Get rid of nonsensical `__getitem__` hook
Summary:
I honestly have no idea what this hook was supposed to do, we
added support for a bare `__getitem__` and a test case that
appears to be literally the only use of it in D19352623; my current
guess is that this was speculative in that we cwere considering some
kind of printable form for overloads that we didn't use.

Since a bare `__getitem__` has no actual meaning in Python (other
than as a runtime method call), I can't see any way that this would
ever be helpful, and it is clearly impossible to map to a subscript
based world.

Just getting rid of it seems best.

Reviewed By: grievejia

Differential Revision: D57002339

fbshipit-source-id: 5286fc5a55408eb2c4fdd845597a26eb04b44d9b