Use explicit `__getitem__`modelQueryExecution updates
Summary:
I'm working toward being able to eliminate the Pyre parser's rewriting
of subscript nodes into `__getitem__` calls.

In order to make that cutover (which is going to inevitably be a *huge*
diff in terms of line counts) sane, I need all of the callsites to be
matching on `__getitem__` explicitly, so that:
  (a) I can find them all with a simple search
  (b) I can do a simple mechanical transform on each one

A problem in the existing code is that quite a few places are relying
on implicit contextual information to *assume* that any call is actually
a subscript - this Pysa code will actually treat `Foo.nonsense(bar)` as
being exactly equivalent to `Foo[bar]`, because it isn't bothering to
verify that the method being called is `__getitem__`.

Making the match explicit is the easy fix.

Reviewed By: alexkassil

Differential Revision: D57011114

fbshipit-source-id: 89cd1b893f290be81b18433b2da5a253ee650ef2