Simplify matching of `__getitem__` in `modelParser`
Summary:
This diff on it's own is probably arguably a step backwards - Pysa is
using some fancy indirection to use slightly more compact code for matching
`__getitem__` calls, and I'm stripping it out to make everything a simple
pattern match.

The reason I want to do this is that I'm looking at cutting Pyre's parser
over to using proper Subscript nodes instead of rewriting subscripts to
`__getitem__` calls.

In the cutover diff, I want all of the callsites to be purely mechanical
swaps of the new subscript with base + index in place of the existing
call of callee = (attribute with base + `__getitem__`) + call argument.

As a result, for my purposes this simpler match logic is better, even if
it looks a little messier in the intervening state; the good news is that
one layer of match boilerplate will disapppear when we cut over so I think
the ugliness introduced here should be temporary and the end state will
be better.

Reviewed By: alexkassil

Differential Revision: D57011632

fbshipit-source-id: 9054a33dfde790bdb0f692a0923407ee24d8ce7d