Handle frozen attributes in dataclass transforms
Summary:
In dataclass_transforms, if we have that frozen_default=True, then we want to mark all attributes as frozen. This diff acomplishes this.

One nuance is that we make attribites and methods ReadOnly but overridable.

We add integration tests that show that we give an error when trying to mutate an attribute that's supposed to be frozen

We remove incorrect unit tests (which I added but realized it does not accomplish the desired results). We will consider how we would refactor those in figure diffs.

In the next diff, we will consider inheritance.

Reviewed By: stroxler, grievejia

Differential Revision: D56862043

fbshipit-source-id: a6cf1743b0fd0128ba7af07cd71e7e6c14148cf8