Rename `Awaitable.create` to `Awaitable.of_location`
Summary:
As someone onboarding to the code, I find it strange to encounter
`Awaitable.create` and not be able to infer that we are "creating
an awaitable" here.

But in fact, *all but one* of these calls is really just making
a lookup key - only *one* of them (the one on line 538 as of this
diff) is actually creating an awaitable that our static analysis
discovered.

Using `of_location` removes the potentially misleading connotation that this
is "creating an awaitable" in a sementic analysis sense and instead is just a
neutral description of the ocaml type algebra which I think is better in this
case.

Reviewed By: grievejia

Differential Revision: D56973177

fbshipit-source-id: 8fddb901394368319a84376ac238c5954a180afa