Extract `forward_setitem` from `forward_call`
Summary:
This commit is the key to understanding how we can close the
regression tracked by D56938609, in which assignments to
a base aren't properly being marked as owned by the base.

The logic we need was buried inside of the `forward_arguments`
helper to `forward_call` up until this commit, which is not possible
to use once we correctly deal with the fact that complex assignment
statements may not be rewriteable as a call expression per se.

My key insight is that *none* of the normal `forward_call` logic
is actually necessary for `__setitem__` as long as the `__setitem__`
behaves sanely in that it returns None. I go ahead and make this
assumption, which allows me to model `__setitem__` *only* in terms
of side effects, and completely disentangle it from `forward_call`.

Now it is possibly that code which is "well typed" under a very narrow
definition of "well-typed" that allows you to do absurd things with
special methods like `__setitem__` could break this assumption, and
the new code would produce a new false positive in that case.

In particular, consider the following:
```
class AbsurdClass:
   def __setitem__(k, v: Awaitable[Whatever]) -> Awaitable[Whatever]:
      ...

output_of_absurd_setitem = absurd_class.__setitem__(k, awaitable_value())

await output_of_absurd_setitem
```

The new logic will ignore the fact that the awaitable owned by
`awaitable_value()` is potentially tracked by `output_of_absurd_setitem`,
and as a result it will complain about an unawaited awaitable whereas
the old code would have not complained.

I don't think this is important: that code is actually *worse* than
ill-typed code because it's breaking the semantics of *dynamic* python,
and in addition "normal" python code doesn't call `__setitem__` directly
anyway - it's always called implicitly via an assignment, in which case
you *can't* propagate values out from it.

So I am pretty strongly in favor of allowing this false positive to
creep in - anyone impacted is abusing Python badly and will just have
to use a `# pyre-ignore` (also I suspect that literally no one is actually
doing this anyway... I did a code search for `=.*__setitem__\(` on a multimillion
line codebase and got 73 hits; of those hits, all the ones I spot-checked were:
- false positives (mostly strings that matched the regex), or
- inside implementations of `__setitem__` that weren't themselves returning values
  but were just handlign the possibility that some *other* class might try to
  (mostly by way of calling `super().__setitem__` and forwarding the output)
- using `__setitem__` inside a lambda to define a setter as a first-class
  function, which is pretty crazy looking but is really just a workaround for
  limitations in python's syntax.

Reviewed By: grievejia

Differential Revision: D56973171

fbshipit-source-id: 643dd4d3601a9ad4c77f9eb0d1c8078267310180