Eliminate `__setitem__` in menhir_parser
Summary:
Even though we'd like to kill the `menhir_parser` code soon, we still
need it for now, for three reasons:
- It's still used in parts of Pyre, especially for expression parsing
  (I am hoping to remove it within H1)
- It's still used by qualification tests, and this might not be worth
  removing now given that we're actively changing qualification and it's
  actually true that having tests written in pseudo-python code makes
  them easier to read and modify
- It's still used for pysa model parsing, and I actually would prefer
  to get the `__getitem__` and `__setitem__` fixes done *before* forking
  so that Pysa won't have `__getitem__` match statements in their model
  parser forever.

This commit updates the parser to use subscripts on the LHS of assignment
instead of trying to rewrite assignments into `__setitem__` calls.

I also took the liberty of eliminating unit tests that don't exist in the
other two parser test suites, such as separately verifying a 5-part assign
instead of just a 3-part and verifying a yield expression / tuple expression
on the RHS. The point of code is to be modular, and we don't need to
exhaustively test all the interactions of various expression-parsing chunks of
logic. This is why the other parsers don't test quite as many combinations
of things, and we don't need to carry around the dead weight in the menhir
parser tests when making AST changes.

Reviewed By: yangdanny97

Differential Revision: D56763428

fbshipit-source-id: b336d84e59821a101b5e3932699afb399540537c