Clarify some docs on unawaitedAwaitableCheck
Summary:
This commit more clearly documents the role of `Awaitable.t`, which
is really just a lookup key (using the location of an awaitable's
"originator" expression as the key) used to map between ownership
vs reachability state

I also tweak some of the top-most comment, in particular to emphasize
that awaiting an awaitable is not the only way to "kill" it in
reachability analysis, we can also transfer ownership to another scope
via `return`, `yield`, or passing it as an argument to a function call.

Note that the vocabulary used here ("transferring ownership to another
scope") is in my view yet more confirmation that "owner" rather than
"alias" is a better name for the concept of "what expression or bound name is
tracking an awaitable right now". The concept is pretty similar to
ownership in Rust, although the analysis is slightly more lax since we
aren't trying to catch duplicate awaits of the same value, whereas in
Rust you have to have *exactly* one owner so you know when to free memory,
which also means we have shared, non-exclusive owners unlike in Rust.

Context:

I'm continuing to work on more properly understanding unawaitedAwaitableCheck
so that I can work out how to fix the bug tracked by D56938609, which
causes a significant regression once I cut the CPython parser over to not
eagerly lowering `__setitem__` calls in D56763430.

Reviewed By: yangdanny97

Differential Revision: D56940226

fbshipit-source-id: dc876cb6e7147fdff17712a923b6f9ee38df1f64