Eliminate `__setitem__` lowering in cpython_parser
Summary:
Instead of trying to convert subscript assignment to `__setitem__` calls,
preserve the subscript (for now, Pyre is lowering this to a `__getitem__`
which we will tackle later).

Also remove the comment claiming that the use of `__getitem__` in a multi-target
assign is wrong, it is in fact correct and the attempt to use `__setitem__`
lowering was wrong in that it could not be fixed for the multi-target case.

**Note**

This is the main cutover diff; I do have diffs making equivalent changes
in the errpy parser and menhir parser, but the normal type checking and pysa
code always rely on the cpython parser as the primary parser.

**Unit test changes (no regressions)**

There are a few analysis-related unit test changes that result:
- of the bugs identified earlier in unawaitedAwaitableCheck that were still
  open after handling multi-assignment are now fixed
- some error messages changed in globalLeakCheckTest (but the actual error
  codes are the same)
- a locationBasedLookupTest case was coupled to the detail that Pyre would
  rewrite the entire span of an assignment into a `__setitem__` call; since
  we no longer do this some of the boundaries change in that test.
- the pysa callgraph changes a bit: a couple of compound callees where
  the `__setitem__` was chaining with something else become simpler. None of
  the actual calls are lost though; none of the locations are dropped, and none
  of the `__setitem__` calls are mistakenly turned into `__getitem__` (which
  happened initially when I wasn't handling `__setitem__` rewrites).

**Pysa behavior**

Pysa is now (locally) rewriting simple subscript assignments back into the old
`__setitem__` form with the logic in D56883262.

As a result, Pysa has no regression on this diff, but Pysa also will continue to be unable to model some cases of multi-target assignment (although it turns out that Pysa's preexisting `__getitem__` behavior for tracking taint on chained assignments like `x[0][1] = 6` causes it to *partially* handle multi-target out-of-the-box... it doesn't handle taint from the index or custom models for `__setitem__` but it does correctly model simple dict-like sementics when keys aren't tainted.

Reviewed By: grievejia

Differential Revision: D56763430

fbshipit-source-id: 96e84e35412917e665b952c2a18368372d349f7b