Refactor logic for marking expression as awaited
Summary:
I suspect that the misleading name of `mark_awaitable_as_awaited`
was a historical accident... it does *not* mark an `Awaitable.t`
as awaited but rather marks a `Location.t` as awaited using two
different paths:
- direct path: if any `Awaitable.t` tracks an awaitable directly
  originating at this location, mark it as awaited
- indirect path: but also mark all `Awaitable.t` awaitables as
  awaited if they are owned by this expression (i.e. if there is
  an `AnonymousExpression` owner at this location).

The actual code path was odd: coerce the location into an `Awaitable.t`, then
unpack the location and cast to an `AnonymousExpression` owner.

This probably made some kind of sense historically, but is kind of absurd
in the current state of the codebase, and also makes the type system
actively mislead us about what's going on (because it looks like we're
keying on originator but we're actually keying on location as either
an originator *or* an AnonymousExpression owner).

As a result, I've renamed the function and refactored so that we just
pass a location. The function name is now reasonable, and we convert
the location into `Awaitable.t` and `AnonymousExpression.t` as
needed. I also have a comment since none of this is going to be obvious
to someone who hasn't looked at `unawaitedAwaitableCheck` before.

Reviewed By: grievejia

Differential Revision: D56941410

fbshipit-source-id: 27451e0a575bebc3268d3faf7ece63483c75bed8