Push `forward_argument` and `is_special_function`
Summary:
In order to fix the regression documented in D56938609, I need
to tease apart the current `__setitem__` handling portion of
`forward_call` until we understand it well enough and have it
modular enough to reproduce the behavior in `forward_assign`
when we see subscripts in the LHS of an assignment.

This process is easier if we avoid closing over irrelevant helper functions in
`__setitem__` logic - it's important to realize that we never call
`forward_argument` or `is_special_function` in the `__setitem__` specialized
branch of `forward_arguments`. This is clearer if we push the definition down
to the other branch.

It's also nicer to make `is_special_function` just data - the callee isn't
changing as we traverse arguments, so it's both inefficient and less
clear what's happening when we keep calling it over and over.

Reviewed By: grievejia

Differential Revision: D56973173

fbshipit-source-id: 8382ea37fe6ea096411ba31cfe1adb5231fe64eb