Change "alias" to "owner" in unawaitedAwaitableCheck
Summary:
I found the use of "alias" here to be misleading - to me an "alias"
implies a "name for a thing" - for example a type alias is a name
for the type it aliases, and an import alias binds a local name to
some other module's global.

The "aliases" in unawaitedAwaitableCheck on the other hand do not
have this property at all:
- they aren't always names, they can just be expressions which is not
  necessarily a problem but does make the name "alias" a little
  questionable
- more importantly, they are many-to-many
  - we actually aggregate awaitables into the same "alias", for example a big
    expression can be an "alias" for many awaitables that pop up in
    sub-expressions, and a named variable can be an "alias" for many awaitables
    in either an assigned expression or that are assigned to subscripts.
  - in addition several "aliases" can track the same awaitable. This isn't
    a problem with the term "alias" since that's normal, it potentially
    does lead to an objection regarding the name "owner", see my note
    below.

All of this convinces me that we're really talking about something
better described as "ownership" here rather than "aliasing". Examples
of where this name is clearer:
- it is not normal vocabulary to say a huge expression "aliases" a half
  dozen sub-expressions, but it makes perfect sense to say it "owns"
  the awaitable state of them.
- it's even less normal vocabulary to say that `d["x"] = foo()` means
  "d is an alias for foo()", but again it makes perfect sense to say
  that "d owns the awaitable state from `foo()`"

In the process of renaming I also expanded the comment on the owner
type and moved it up - I find it is sometimes more helpful to document the
purpose of a type as a whole and describe how the variants relate to
one another in a single block rather than documenting the variants
in separate blocks, and in my opinion with ownership of awaitables this
is the easier way to explain what is happening.

Moreover, in the top-level comment I tweak the wording to stop saying
we are doing an Alias analysis, which I don't think is exactly correct;
alias analysis refers to scenarios where a compiler determines whether two
pointers can point to the same value
(https://en.wikipedia.org/wiki/Alias_analysis), which respects my intuition
that "alias" should mean two values might be in some sense equivalent.

In in our case we're asking about *reachability* via pointer traversals, which
I've described as "ownership" or "aliasability" (reachability isn't a great
name because that's already being used to describe dataflow as opposed to
pointer traversal in the context of this analysis).

NOTE: the name `owner` is debateable here. I'm very sure it's better than
`alias`, and in fact the concept is pretty similar to `owner` in Rust. But it
isn't the same as Rust - unawaitedAwaitableCheck owners are *shared* owners,
so a single awaitable can be tracked in more than one (the relationship
is in general many-to-many). I've emphasized this and provided an example
in the new comment block, but this is also a reason why we might want a
different name. I considered `holder` as a reasonable name for "shared owner",
or `referee` since the point is it's a "thing from which there might be a chain of
references to the awaitable" but decided that `owner` is more familiar - happy
to change this if someone prefers one of the other options or has a better name.

NOTE: I remain a little confused as to exactly why we need both
AnonymousExpression owners for chained method calls *and*
`nested_aliased_expressions` in the `forward_expression_result` type;
there's something surprising going on here that isn't yet very well explained - why
can't we just mark all of the expressions inside `nested_aliased_expressions` as
awaited if we see an `await` on a chaned method call? I tried to preserve the
existing comments (which just say "we need it") but I don't feel like this has
enough explanation yet.

Reviewed By: grievejia

Differential Revision: D56938610

fbshipit-source-id: 8491f1d4ced213ff2cb6eda67a747f080946bf01