Fix a gap in key analysis for (existing) `__setitem__`
Summary:
The preexisting `__setitem__` logic in unawaitedAwaitableCheck
can lead to both false positives *and* false negatives when analyzing
the update key matters for the analysis.

In practice it's probably unusual to do awaitable-relevant logic in a
key; in particular using awaitable *values* as index keys is probably
highly abnormal (you'd have to implement `__hash__` on an awaitable
type). But using `await` in an index is more plausible and we would
have a false positive here prior to this diff.

Fixing this gap before we try to fix the regression in D56938609 is
important since we want our eventual solution to model the control flow
correctly in all cases, not just the "typical" case where key expressions
aren't relevant to awaitable analysis.

Reviewed By: grievejia

Differential Revision: D56973175

fbshipit-source-id: 96856c8ab385ad825c5e362e74d3213205750196