Tweak test cases
Summary:
It turns out that my test case added in D56938609 to track a regression that
happens when I try to eliminate `__setitem__` lowering from the cpython parser
isn't as useful as I wanted.

The problem is that there's a known gap in the analysis (see the comment on line 959
of `unawaitedAwaitableCheck.ml` as of this commit) where we don't propagate
ownership of awaitables when we cannot destructure a literal tuple on the LHS
along with the RHS.

Since that issue is pre-existing and not tied directly to the parser changes that
I'm trying to make, I don't actually want to focus on a test that tracks this.

As a result, I split that test into two tests. The one marked with
`TODO(T101303314)` now has a literal tuple on the RHS, and a fully-correct fix
for `__setitem__` lowering ought to fix it. The other one, on the other hand, is
a pre-existing limitation in the analysis and I don't expect to fix it anytime
soon.

Reviewed By: grievejia

Differential Revision: D56973176

fbshipit-source-id: 7ede937bf251dcce70a7fee79f17cc01c84effb5