Rewrite __getitem__ assignments into __setitem__ calls for pysa
Summary:
In a follow-up (D56763430), we will stop turning `d[a] = b` into `__setitem__` calls during AST parsing.
Unfortunately, this would lead to false negatives in Pysa, which is relying on this behavior.
For now, let's perform the rewriting of the AST directly in Pysa to preserve the old behavior, until we can handle subscripts assignments properly.

Reviewed By: arthaud

Differential Revision: D56883262

fbshipit-source-id: b41d3679bb5c72983c768acd826fa55e4114268f