Hide details of resolve alias from tests
Summary:
For reasons that I'll explain more in the next diff, the existing
"resolve callable getitem" logic is an implementation detail that has
to break when handling subscripts properly in the AST, and therefore
does not belong in the test suite.

The underlying capability of handling type aliases for Callable *does* deserve
testing, but the tests should rely on `Type.create`, not in a helper function
whose implementation is brittle with respect to AST representations.

This commit keeps the same test logic (except for the trivial cases that
are covered by other tests) but uses only `Type.create` to power them.

Reviewed By: samwgoldman

Differential Revision: D57072628

fbshipit-source-id: 037035b8bcf0f3a224658343efa3c76dbdb3ec74