Make it possible to isolate Type tests
Summary:
The `Type.ml` codebase is huge and very hard to work on, plus the test
suite is also huge.

It looks like at some point we attempted to break the test suite into
smaller groupings, but the way we were doing it is wrong - we were
using multiple calls to the `Test.run` function at the top level, which
makes it impossible to list tests or run a single test using OUNIT. This
is never okay to do, Pyre cannot be print debugged at all without some
ability to isolate test cases.

Reviewed By: samwgoldman

Differential Revision: D57072626

fbshipit-source-id: b2ab0485d68b801e9a71167da0a3abc2fba8795b