Extract the `__getitem__` match in expression-to-type logic
Summary:
Prior to this diff, we were using a very fat top-level
branch in the expression-to-type conversion logic that handled
all kinds of cases of `__getitem__`.

This diff changes things so that there's a single top-level branch
and we match on the base + subscript separately.

I'm not sure this is a huge win for "general" readability. Fewer top-level
branching is nice but multiple levels of matches are annoying, and it's
probably about the same overall.

But this factoring makes it significantly easier for me to do the specific
refactor I need right now, which is working out which helper functions have the
wrong type signatures. The reason is that all of the problems get consolidated
into a single match, and I will know I'm ready to change the parser when I no longer need
to use `callee_value`, `getitem_arguments`, or `getitem_call` (none of which
will exist anymore once we switch to subscripts: we will only have `base` and
`index_expression` to work with).

Reviewed By: samwgoldman, grievejia

Differential Revision: D57072631

fbshipit-source-id: a5bac5865be200499987803e4b4671759cf52e21