Cleanup create_logic: unify names
Summary:
At this point the `Type.create_logic` function works
in a way that we can cut over to producing an `Expression.Subscript`
node from the parser. But along the way, a few things got messy
that I'd like to clean up.

This cleanup diff unifies the names for a few things across helpers,
reducing the number of renames needed and the amount of context switching
needed when looking across several helpers (anytime we use inconsistent
names, it makes comparing blocks of code with different names harder)

Reviewed By: grievejia

Differential Revision: D57072630

fbshipit-source-id: 5c3ade44191c1d056902c7b96659c8e347f58028