Make `__getitem__` explicit in the general `Call` case
Summary:
**This stack: simplify Callable type (and all generic) parsing**

I realized when looking at `Type.ml` that the current logic cannot
be changed in place to handle getting rid of `__getitem__` lowering
in the parser, because the callable type parsing has been made
somewhat complex to handle directly representing both named callable
types and overloads in tests.

We could get rid of this, but there are actually quite a lot of tests
that would have to change, so in the interests of keeping my AST
improvements going I think it's better for now to just improve the code
than to rewrite dozens of tests.

My goal is to rewrite the existing logic (with no observable behaviors;
tests shouldn't change at all in this stack) until all the subscript
matching is done in explicit match statements that involve `__getitem__`,
because then *and only then* will it be a simple turn-the-crank refactor
to cut over to a new `Expresison.Subscript` node.

It turns out this is pretty involved because incorrect (once we model the
actual CPython AST) assumptions about type signatures are being used
many layers down the call stack.

As a result, the fix will be a medium-size stack that alternates pushing
blocks of code into new places with tweaking the type signatures (and
on occasion fixing tests that aren't compatible with the direction we
want to go); most of the diffs are pretty simple individually but it's a
lot of code being changed overall.

**This diff: make it clear that *all* top-level Calls are subscripts**

The problem I want to solve here is that the current combination
of `is_typing_callable` and `parse_callable` cannot be cut over
to subscript nodes without refactors.

It's going to be a little tricky to solve this because both functions
are recursively matching chains of `__getitem__` and other calls
in order to handle Pyre-only extensions of Callable like
```
typing.Callable("myclass.mymethod")[[arg], return][
  typing.Callable[[overload_arg_0], overload_return_0],
  typing.Callable[[overload_arg_0], overload_return_1],
]
```

But the current state of the code is such that you can't even be
sure the *top-level* expression is a subscript. It actually is,
and making that explicit in this diff makes it easier to wrap our
head around what happens in lower-layer code. Note that a `Call`
which is not either `__getitem__` or one of the earlier cases
handled by a more direct match will now fall through to the final
`Top` catch-all on line 4344; the earlier code was producing `Top`
too but was just much less obvious about it (for example,
`parse_callable` returns `Top` if it sees a form it can't match).

Reviewed By: grievejia

Differential Revision: D57072635

fbshipit-source-id: d1baa543d20789c89b13cd87a4119b81df975a55